Move create_refactor to create
