+++ pkg/odo/cli/component/create.go
@@ -3,33 +3,28 @@ package component
 import (
 	"fmt"
 	"io/ioutil"
-	"os"
 	"path/filepath"
 	"strings"
 
-	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
-	"github.com/zalando/go-keyring"
-
-	registryLibrary "github.com/devfile/registry-support/registry-library/library"
-	"github.com/openshift/odo/pkg/catalog"
-	"github.com/openshift/odo/pkg/component"
-	"github.com/openshift/odo/pkg/devfile"
+	"github.com/devfile/library/pkg/devfile"
+	"github.com/devfile/library/pkg/devfile/parser"
+	odoDevfile "github.com/openshift/odo/pkg/devfile"
 	"github.com/openshift/odo/pkg/devfile/location"
-	"github.com/openshift/odo/pkg/devfile/validate"
 	"github.com/openshift/odo/pkg/envinfo"
 	"github.com/openshift/odo/pkg/log"
-	appCmd "github.com/openshift/odo/pkg/odo/cli/application"
-	"github.com/openshift/odo/pkg/odo/cli/component/ui"
-	projectCmd "github.com/openshift/odo/pkg/odo/cli/project"
-	registryConsts "github.com/openshift/odo/pkg/odo/cli/registry/consts"
 	registryUtil "github.com/openshift/odo/pkg/odo/cli/registry/util"
+	scontext "github.com/openshift/odo/pkg/segment/context"
+	"github.com/openshift/odo/pkg/util"
+	"github.com/pkg/errors"
+	"github.com/zalando/go-keyring"
+
+	"github.com/openshift/odo/pkg/catalog"
+	appCmd "github.com/openshift/odo/pkg/odo/cli/application"
+	projectCmd "github.com/openshift/odo/pkg/odo/cli/project"
 	"github.com/openshift/odo/pkg/odo/genericclioptions"
 	odoutil "github.com/openshift/odo/pkg/odo/util"
 	"github.com/openshift/odo/pkg/odo/util/completion"
-	"github.com/openshift/odo/pkg/preference"
-	scontext "github.com/openshift/odo/pkg/segment/context"
-	"github.com/openshift/odo/pkg/util"
+	"github.com/spf13/cobra"
 
 	ktemplates "k8s.io/kubectl/pkg/util/templates"
 )
@@ -118,325 +113,139 @@ func NewCreateOptions() *CreateOptions {
 	}
 }
 
-func createDefaultComponentName(componentType string, sourcePath string) (string, error) {
-	finalSourcePath, err := getComponentName(sourcePath)
+func (co *CreateOptions) Complete(name string, cmd *cobra.Command, args []string) (err error) {
+	// GETTERS
+	// Get context
+	co.Context, err = getContext(co.now, cmd)
 	if err != nil {
-		return "", err
+		return err
 	}
-
-	return component.GetDefaultComponentName(
-		finalSourcePath,
-		componentType,
-		component.ComponentList{},
-	)
-}
-
-func getComponentName(sourcePath string) (string, error) {
-	// Retrieve the componentName, if the componentName isn't specified, we will use the default image name
-	// we only get absolute path for local source type
-	if sourcePath == "" {
-		wd, err := os.Getwd()
-		if err != nil {
-			return "", err
-		}
-		return wd, nil
-	}
-	finalSourcePath, err := filepath.Abs(sourcePath)
-	if err != nil {
-		return "", err
-	}
-	return finalSourcePath, nil
-
-}
-
-// Complete completes create args
-func (co *CreateOptions) Complete1(name string, cmd *cobra.Command, args []string) (err error) {
-
-	if co.now {
-		// this populates the EnvInfo as well
-		co.Context, err = genericclioptions.NewContextCreatingAppIfNeeded(cmd)
-		if err != nil {
-			return err
-		}
-	} else {
-		co.Context, err = genericclioptions.NewOfflineContext(cmd)
-		if err != nil {
-			return err
-		}
-	}
-
-	DevfilePath := location.DevfileLocation("")
-	envFilePath := ""
-	// Configure the context
-	if co.componentContext != "" {
-		DevfilePath = location.DevfileLocation(co.componentContext)
-		envFilePath = filepath.Join(co.componentContext, EnvYAMLFilePath)
-		co.PushOptions.componentContext = co.componentContext
-	}
-	co.DevfilePath = DevfilePath
-
-	if util.CheckPathExists(envFilePath) && util.CheckPathExists(co.DevfilePath) {
-		return errors.New("this directory already contains a component")
-	}
-
-	if util.CheckPathExists(envFilePath) && !util.CheckPathExists(co.DevfilePath) {
-		log.Warningf("Found a dangling env file without a devfile, overwriting it")
-		if err = util.DeletePath(envFilePath); err != nil {
-			return err
-		}
-	}
-
-	if util.CheckPathExists(co.DevfilePath) && co.devfileMetadata.devfilePath.value != "" && !util.PathEqual(co.DevfilePath, co.devfileMetadata.devfilePath.value) {
-		return errors.New("this directory already contains a devfile, you can't specify devfile via --devfile")
-	}
-
-	// we check if the devfile is already present or not, this location is important - check should happen early
-	if util.CheckPathExists(co.DevfilePath) {
-		co.devfileMetadata.userCreatedDevfile = true
-	}
-
+	// Get the app name
 	co.appName = genericclioptions.ResolveAppFlag(cmd)
-
-	isDevfileRegistryPresent := true // defaulted to true since odo ships with a default registry set
-	var catalogDevfileList catalog.DevfileComponentTypeList
-
-	// Validate user specify devfile path
-	if co.devfileMetadata.devfilePath.value != "" {
-		fileErr := util.ValidateFile(co.devfileMetadata.devfilePath.value)
-		urlErr := util.ValidateURL(co.devfileMetadata.devfilePath.value)
-		if fileErr != nil && urlErr != nil {
-			return errors.Errorf("the devfile path you specify is invalid with either file error \"%v\" or url error \"%v\"", fileErr, urlErr)
-		} else if fileErr == nil {
-			co.devfileMetadata.devfilePath.protocol = "file"
-		} else if urlErr == nil {
-			co.devfileMetadata.devfilePath.protocol = "http(s)"
-		}
-	}
-
-	// Validate user specify registry
-	if co.devfileMetadata.devfileRegistry.Name != "" {
-		if co.devfileMetadata.devfilePath.value != "" {
-			return errors.New("you can't specify registry via --registry if you want to use the devfile that is specified via --devfile")
-		}
-
-		registryList, e := catalog.GetDevfileRegistries(co.devfileMetadata.devfileRegistry.Name)
-		if e != nil {
-			return errors.Wrap(e, "failed to get registry")
-		}
-		if len(registryList) == 0 {
-			return errors.Errorf("registry %s doesn't exist, please specify a valid registry via --registry", co.devfileMetadata.devfileRegistry.Name)
-		}
-	}
-
-	// Can't use the existing devfile or download devfile from registry, go to interactive mode
+	// Get the project name
+	co.devfileMetadata.componentNamespace = co.Context.Project
+	// Get DevfilePath
+	co.DevfilePath = location.DevfileLocation(co.componentContext)
+	//Check whether the directory already contains a devfile, this check should happen early
+	co.devfileMetadata.userCreatedDevfile = util.CheckPathExists(co.DevfilePath)
+	// EnvFilePath is the path of env file for devfile component
+	envFilePath := getEnvFilePath(co.componentContext)
+	// This is required so that .odo is created in the correct context
+	co.PushOptions.componentContext = co.componentContext
+	// Use Interactive mode if: 1) no args are passed || 2) the devfile exists || 3) --devfile is used
 	if len(args) == 0 && !util.CheckPathExists(co.DevfilePath) && co.devfileMetadata.devfilePath.value == "" {
 		co.interactive = true
 	}
 
-	var componentType string
-	var componentName string
-	var componentNamespace string
-
-	if co.interactive {
-		// Interactive mode
-
-		// Get available devfile components for checking devfile compatibility
-		catalogDevfileList, err = catalog.ListDevfileComponents(co.devfileMetadata.devfileRegistry.Name)
+	// CONFLICT CHECK
+	// Check if a component exists
+	if util.CheckPathExists(envFilePath) && util.CheckPathExists(co.DevfilePath) {
+		return errors.New("this directory already contains a component")
+	}
+	// Check if there is a dangling env file; delete the env file if found
+	if util.CheckPathExists(envFilePath) && !util.CheckPathExists(co.DevfilePath) {
+		log.Warningf("Found a dangling env file without a devfile, overwriting it")
+		// Note: if the IF condition seems to have a side-effect, it is better to do the condition check separately, like below
+		err = util.DeletePath(envFilePath)
 		if err != nil {
 			return err
 		}
-
-		if len(catalogDevfileList.DevfileRegistries) == 0 {
-			isDevfileRegistryPresent = false
-			log.Warning("Registry is empty, please run `odo registry add <registry name> <registry URL>` to add a registry\n")
-		}
-
-		if isDevfileRegistryPresent {
-			// Component type: We provide devfile component list to let user choose
-			componentType = ui.SelectDevfileComponentType(catalogDevfileList.Items)
-
-			// Component name: User needs to specify the component name, by default it is component type that user chooses
-			componentName = ui.EnterDevfileComponentName(componentType)
-
-			// Component namespace: User needs to specify component namespace, by default it is the current active namespace
-			if cmd.Flags().Changed("project") {
-				componentNamespace, err = cmd.Flags().GetString("project")
-				if err != nil {
-					return err
-				}
-			} else {
-				client, e := genericclioptions.Client()
-				// if the user is logged in or if we have cluster information, display the default project
-				if e == nil {
-					componentNamespace = ui.EnterDevfileComponentProject(client.GetCurrentProjectName())
-				}
-			}
-		}
-
-	} else {
-		// Direct mode (User enters the full command)
-
-		if util.CheckPathExists(co.DevfilePath) || co.devfileMetadata.devfilePath.value != "" {
-			// Use existing devfile directly
-
-			if len(args) > 1 {
-				return errors.Errorf("accepts between 0 and 1 arg when using existing devfile, received %d", len(args))
-			}
-
-			// If user can use existing devfile directly, the first arg is component name instead of component type
-			if len(args) == 1 {
-				componentName = args[0]
-			} else {
-				// If there is an existing devfile, and no component name is passed, parse it from the devfile,
-				// and assign the value if the metadata name is set
-				devfileObj, e := devfile.ParseAndValidateFromFile(DevfilePath)
-				if e == nil && devfileObj.GetMetadataName() != "" {
-					componentName = devfileObj.GetMetadataName()
-				} else {
-					// If the metadata name is not available, then assign the current directory name to component
-					currentDirPath, wdErr := os.Getwd()
-					if wdErr != nil {
-						return wdErr
-					}
-					currentDirName := filepath.Base(currentDirPath)
-					componentName = currentDirName
-				}
-			}
-
-			co.devfileMetadata.devfileSupport = true
-		} else if len(args) > 0 {
-			// Download devfile from registry
-
-			// Component type: Get from full command's first argument (mandatory in direct mode)
-			componentType = args[0]
-
-			// Component name: Get from full command's second argument (optional in direct mode), by default it is a generated name if second arg is not provided
-			if len(args) == 2 {
-				componentName = args[1]
-			} else {
-				componentName, err = createDefaultComponentName(
-					componentType,
-					co.componentContext,
-				)
-				if err != nil {
-					return err
-				}
-			}
-
-			// Get available devfile components for checking devfile compatibility
-			catalogDevfileList, err = catalog.ListDevfileComponents(co.devfileMetadata.devfileRegistry.Name)
-			if err != nil {
-				return err
-			}
-			if co.devfileMetadata.devfileRegistry.Name != "" && catalogDevfileList.Items == nil {
-				return errors.Errorf("can't create devfile component from registry %s", co.devfileMetadata.devfileRegistry.Name)
-			}
-
-			if len(catalogDevfileList.DevfileRegistries) == 0 {
-				isDevfileRegistryPresent = false
-				log.Warning("Registry list is empty, please run `odo registry add <registry name> <registry URL>` to add a registry\n")
-			}
-		}
-
-		componentNamespace = co.Context.Project
+	}
+	//Check if the directory already contains a devfile when --devfile flag is passed
+	if util.CheckPathExists(co.DevfilePath) && co.devfileMetadata.devfilePath.value != "" && !util.PathEqual(co.DevfilePath, co.devfileMetadata.devfilePath.value) {
+		return errors.New("this directory already contains a devfile, you can't specify devfile via --devfile")
+	}
+	// Check if both --devfile and --registry flag are used, in which case raise an error
+	if co.devfileMetadata.devfileRegistry.Name != "" && co.devfileMetadata.devfilePath.value != "" {
+		return errors.New("you can't specify registry via --registry if you want to use the devfile that is specified via --devfile")
+	}
+	// More than one arguments should not be allowed when a devfile exists or --devfile is used
+	if len(args) > 1 && (co.devfileMetadata.userCreatedDevfile || co.devfileMetadata.devfilePath.value != "") {
+		return errors.Errorf("accepts between 0 and 1 arg when using existing devfile, received %d", len(args))
 	}
 
-	// set devfileName to same value as componentType for telemetry
-	co.devfileName = componentType
+	// Initialize envinfo
+	err = co.InitEnvInfoFromContext()
+	if err != nil {
+		return err
+	}
+
+	// Set the starter project token if required
+	if co.devfileMetadata.starter != "" {
+		var secure bool
+		secure, err = registryUtil.IsSecure(co.devfileMetadata.devfileRegistry.Name)
+		if err != nil {
+			return err
+		}
+		if co.devfileMetadata.starterToken == "" && secure {
+			var token string
+			token, err = keyring.Get(fmt.Sprintf("%s%s", util.CredentialPrefix, co.devfileMetadata.devfileRegistry.Name), registryUtil.RegistryUser)
+			if err != nil {
+				return errors.Wrap(err, "unable to get secure registry credential from keyring")
+			}
+			co.devfileMetadata.starterToken = token
+		}
+	}
+
+	log.Info("Devfile Object Creation")
+	var catalogDevfileList catalog.DevfileComponentTypeList
+	var createMethod CreateMethod
+	if co.devfileMetadata.devfilePath.value != "" {
+		fileErr := util.ValidateFile(co.devfileMetadata.devfilePath.value)
+		urlErr := util.ValidateURL(co.devfileMetadata.devfilePath.value)
+		if fileErr != nil && urlErr != nil {
+			return errors.Errorf("the devfile path you specify is invalid with either file error %q or url error %q", fileErr, urlErr)
+		} else if fileErr == nil {
+			co.devfileMetadata.devfilePath.protocol = "file"
+			createMethod = FileCreateMethod{}
+
+		} else if urlErr == nil {
+			co.devfileMetadata.devfilePath.protocol = "http(s)"
+			createMethod = HTTPCreateMethod{}
+		}
+	}
+	switch {
+	case co.devfileMetadata.devfilePath.value != "" || co.devfileMetadata.userCreatedDevfile:
+		//co.devfileName = "" for user provided devfile
+		err = createMethod.FetchDevfileAndCreateComponent(co, catalogDevfileList)
+		if err != nil {
+			createMethod.Rollback(co.DevfilePath)
+			return err
+		}
+		err = createMethod.SetMetadata(co, cmd, args, catalogDevfileList)
+		if err != nil {
+			createMethod.Rollback(co.DevfilePath)
+			return err
+		}
+	default:
+		if co.interactive {
+			createMethod = InteractiveCreateMethod{}
+		} else {
+			createMethod = DirectCreateMethod{}
+		}
+
+		catalogDevfileList, err = validateAndFetchRegistry(co.devfileMetadata.devfileRegistry.Name)
+		if err != nil {
+			return err
+		}
+		err = createMethod.SetMetadata(co, cmd, args, catalogDevfileList)
+		if err != nil {
+			return err
+		}
+		err = createMethod.FetchDevfileAndCreateComponent(co, catalogDevfileList)
+		if err != nil {
+			createMethod.Rollback(co.DevfilePath)
+			return err
+		}
+	}
+	// Adding user provided devfile name to telemetry data
 	scontext.SetDevfileName(cmd.Context(), co.devfileName)
 
-	// Set devfileMetadata struct
-	co.devfileMetadata.componentType = componentType
-	co.devfileMetadata.componentName = strings.ToLower(componentName)
-	co.devfileMetadata.componentNamespace = strings.ToLower(componentNamespace)
-
-	if util.CheckPathExists(co.DevfilePath) || co.devfileMetadata.devfilePath.value != "" {
-		// Categorize the sections
-		log.Info("Devfile Object Validation")
-
-		var devfileAbsolutePath string
-		if util.CheckPathExists(co.DevfilePath) || co.devfileMetadata.devfilePath.protocol == "file" {
-			var devfilePath string
-			if util.CheckPathExists(co.DevfilePath) {
-				devfilePath = co.DevfilePath
-			} else {
-				devfilePath = co.devfileMetadata.devfilePath.value
-			}
-			devfileAbsolutePath, err = filepath.Abs(devfilePath)
-			if err != nil {
-				return err
-			}
-		} else if co.devfileMetadata.devfilePath.protocol == "http(s)" {
-			devfileAbsolutePath = co.devfileMetadata.devfilePath.value
-		}
-		devfileSpinner := log.Spinnerf("Creating a devfile component from devfile path: %s", devfileAbsolutePath)
-		defer devfileSpinner.End(true)
-
-		// Initialize envinfo
-		err = co.InitEnvInfoFromContext()
-		if err != nil {
-			return err
-		}
-		return nil
-	}
-
-	// TODO: Put this inside GetDevfileFromRegistry
-	if isDevfileRegistryPresent {
-		// Categorize the sections
-
-		// Since we need to support both devfile and s2i, so we have to check if the component type is
-		// supported by devfile, if it is supported we return and will download the corresponding devfile later,
-		// if it is not supported we still need to run all the codes related with s2i after devfile compatibility check
-
-		hasComponent := false
-		var devfileExistSpinner *log.Status
-		log.Info("Devfile Object Validation")
-		devfileExistSpinner = log.Spinner("Checking devfile existence")
-		defer devfileExistSpinner.End(false)
-
-		for _, devfileComponent := range catalogDevfileList.Items {
-			if co.devfileMetadata.componentType == devfileComponent.Name {
-				hasComponent = true
-				co.devfileMetadata.devfileSupport = true
-				co.devfileMetadata.devfileLink = devfileComponent.Link
-				co.devfileMetadata.devfileRegistry = devfileComponent.Registry
-				break
-			}
-		}
-
-		if hasComponent {
-			devfileExistSpinner.End(true)
-		} else {
-			devfileExistSpinner.End(false)
-		}
-
-		if co.devfileMetadata.devfileSupport {
-			registrySpinner := log.Spinnerf("Creating a devfile component from registry: %s", co.devfileMetadata.devfileRegistry.Name)
-			if registryUtil.IsGitBasedRegistry(co.devfileMetadata.devfileRegistry.URL) {
-				registryUtil.PrintGitRegistryDeprecationWarning()
-			}
-			// Initialize envinfo
-			err = co.InitEnvInfoFromContext()
-			if err != nil {
-				registrySpinner.End(false)
-				return err
-			}
-
-			registrySpinner.End(true)
-			return nil
-		}
-
-		return fmt.Errorf("devfile component type %q is not supported, please run `odo catalog list components` for a list of supported devfile component types", co.devfileMetadata.componentType)
-
-	}
-	return
+	return nil
 }
 
-// Validate validates the create parameters
-func (co *CreateOptions) Validate1() (err error) {
-
+func (co *CreateOptions) Validate() (err error) {
 	log.Info("Validation")
-
 	// Validate if the devfile component name that user wants to create adheres to the k8s naming convention
 	spinner := log.Spinner("Validating if devfile name is correct")
 	defer spinner.End(false)
@@ -445,154 +254,57 @@ func (co *CreateOptions) Validate1() (err error) {
 	if err != nil {
 		return err
 	}
+	spinner.End(true)
 
+	// Validate if the devfile is compatible with odo; this checks the resolved/flattened version of devfile
+	spinner = log.Spinner("Validating the devfile for odo")
+	defer spinner.End(false)
+
+	_, err = odoDevfile.ParseAndValidateFromFile(co.DevfilePath)
+	if err != nil {
+		return err
+	}
 	spinner.End(true)
 
 	return nil
 }
 
-// Run has the logic to perform the required actions as part of command
-func (co *CreateOptions) devfileRun(cmd *cobra.Command) (err error) {
-	var devfileData []byte
-	devfileExist := util.CheckPathExists(co.DevfilePath)
-	// Use existing devfile directly from --devfile flag
-	if co.devfileMetadata.devfilePath.value != "" {
-		if co.devfileMetadata.devfilePath.protocol == "http(s)" {
-			// User specify devfile path is http(s) URL
-			params := util.HTTPRequestParams{
-				URL:   co.devfileMetadata.devfilePath.value,
-				Token: co.devfileMetadata.token,
-			}
-			devfileData, err = util.DownloadFileInMemory(params)
-			if err != nil {
-				return errors.Wrapf(err, "failed to download devfile for devfile component from %s", co.devfileMetadata.devfilePath.value)
-			}
-		} else if co.devfileMetadata.devfilePath.protocol == "file" {
-			devfileData, err = ioutil.ReadFile(co.devfileMetadata.devfilePath.value)
-			if err != nil {
-				return errors.Wrapf(err, "failed to read devfile from %s", co.devfileMetadata.devfilePath)
-			}
-		}
-	} else {
-		if devfileExist {
-			// if local devfile already exists read that
-			// odo create command was expected in a directory already containing devfile
-			devfileData, err = ioutil.ReadFile(co.DevfilePath)
-			if err != nil {
-				return errors.Wrapf(err, "failed to read devfile from %s", co.DevfilePath)
-			}
-		} else {
-			// Download devfile from registry
-			var params util.HTTPRequestParams
+func (co *CreateOptions) Run(cmd *cobra.Command) (err error) {
+	// Adding component type to telemetry data
+	scontext.SetComponentType(cmd.Context(), co.devfileMetadata.componentType)
 
-			if strings.Contains(co.devfileMetadata.devfileRegistry.URL, "github") {
-				// Github-based registry
-				params = util.HTTPRequestParams{
-					URL: co.devfileMetadata.devfileRegistry.URL + co.devfileMetadata.devfileLink,
-				}
-
-				secure, e := registryUtil.IsSecure(co.devfileMetadata.devfileRegistry.Name)
-				if e != nil {
-					return e
-				}
-
-				if secure {
-					var token string
-					token, err = keyring.Get(fmt.Sprintf("%s%s", util.CredentialPrefix, co.devfileMetadata.devfileRegistry.Name), registryUtil.RegistryUser)
-					if err != nil {
-						return errors.Wrap(err, "unable to get secure registry credential from keyring")
-					}
-					params.Token = token
-				}
-			} else {
-				err = registryLibrary.PullStackFromRegistry(co.devfileMetadata.devfileRegistry.URL, co.devfileMetadata.componentType, co.componentContext, false, registryConsts.TelemetryClient)
-				if err != nil {
-					return err
-				}
-				// if the function fails, remove this newly created devfile
-				defer func() {
-					if err != nil {
-						os.Remove(co.DevfilePath)
-					}
-				}()
-			}
-
-			var cfg *preference.PreferenceInfo
-			cfg, err = preference.New()
-			if err != nil {
-				return err
-			}
-
-			if strings.Contains(co.devfileMetadata.devfileRegistry.URL, "github") {
-				devfileData, err = util.DownloadFileInMemoryWithCache(params, cfg.GetRegistryCacheTime())
-			} else {
-				devfileData, err = ioutil.ReadFile(co.DevfilePath)
-			}
-			if err != nil {
-				return errors.Wrapf(err, "failed to download devfile for devfile component from %s", co.devfileMetadata.devfileRegistry.URL+co.devfileMetadata.devfileLink)
-			}
-		}
-	}
-	// TODO: this should be replaced with github.com/openshift/odo/pkg/devile.ParseAndValidateFromFile(DevfilePath)
-	// But this can be only after we deprecate support for "github based" registries.
-	// When we do that the above "if" will be deleted and parsing from []data won't be necessary
-	devObj, err := devfile.ParseAndValidateFromData(devfileData)
+	devObj, err := devfileParseFromFile(co.DevfilePath, false)
 	if err != nil {
-		return errors.Wrap(err, "unable to parse devfile")
+		return errors.New("Failed to parse the devfile")
 	}
-	// Add component type in case it is not already added or is empty
-	if scontext.GetTelemetryStatus(cmd.Context()) {
-		if value, ok := scontext.GetContextProperties(cmd.Context())[scontext.ComponentType]; !ok || value == "" {
-			scontext.SetComponentType(cmd.Context(), component.GetComponentTypeFromDevfileMetadata(devObj.Data.GetMetadata()))
-		}
-	}
-	err = validate.ValidateDevfileData(devObj.Data)
+
+	devfileData, err := ioutil.ReadFile(co.DevfilePath)
 	if err != nil {
 		return err
 	}
-
-	secure, err := registryUtil.IsSecure(co.devfileMetadata.devfileRegistry.Name)
-	if err != nil {
-		return err
-	}
-	if co.devfileMetadata.starterToken == "" && secure {
-		var token string
-		token, err = keyring.Get(fmt.Sprintf("%s%s", util.CredentialPrefix, co.devfileMetadata.devfileRegistry.Name), registryUtil.RegistryUser)
-		if err != nil {
-			return errors.Wrap(err, "unable to get secure registry credential from keyring")
-		}
-		co.devfileMetadata.starterToken = token
-	}
-
+	// WARN: Starter Project uses go-git that overrides the directory content, there by deleting the existing devfile.
 	err = decideAndDownloadStarterProject(devObj, co.devfileMetadata.starter, co.devfileMetadata.starterToken, co.interactive, co.componentContext)
 	if err != nil {
 		return errors.Wrap(err, "failed to download project for devfile component")
 	}
 
-	// save devfile and corresponding resources if possible
-	// use original devfileData to persist original formatting of the devfile file
+	// TODO: We should not have to rewrite to the file. Fix the starter project.
 	err = ioutil.WriteFile(co.DevfilePath, devfileData, 0644) // #nosec G306
 	if err != nil {
-		return errors.Wrapf(err, "unable to save devfile to %s", co.DevfilePath)
-	}
-	if co.devfileMetadata.devfilePath.value == "" && !devfileExist && !strings.Contains(co.devfileMetadata.devfileRegistry.URL, "github") {
-		err = registryLibrary.PullStackFromRegistry(co.devfileMetadata.devfileRegistry.URL, co.devfileMetadata.componentType, co.componentContext, false, registryConsts.TelemetryClient)
-		if err != nil {
-			return err
-		}
+		return err
 	}
 
-	// set user provided component name in the devfile
-	if co.devfileMetadata.componentName != "" {
-		devObj, err = devfile.ParseAndValidateFromFile(co.DevfilePath)
-		if err != nil {
-			return fmt.Errorf("failed to create devfile component: %w", err)
-		}
+	// If user provided a custom name, re-write the devfile
+	// ENSURE: co.devfileMetadata.componentName != ""
+	if co.devfileMetadata.componentName != devObj.GetMetadataName() {
+		spinner := log.Spinnerf("Updating the devfile with component name %q", co.devfileMetadata.componentName)
+		defer spinner.End(false)
 
 		err = devObj.SetMetadataName(co.devfileMetadata.componentName)
 		if err != nil {
-			return fmt.Errorf("failed to create devfile component: %w", err)
+			return errors.New("Failed to update the devfile")
 		}
+		spinner.End(true)
 	}
 
 	// Generate env file
@@ -629,22 +341,11 @@ func (co *CreateOptions) devfileRun(cmd *cobra.Command) (err error) {
 	} else {
 		log.Italic("\nPlease use `odo push` command to create the component with source deployed")
 	}
-	return nil
-}
 
-// Run has the logic to perform the required actions as part of command
-func (co *CreateOptions) Run1(cmd *cobra.Command) (err error) {
-
-	if scontext.GetTelemetryStatus(cmd.Context()) {
-		scontext.SetComponentType(cmd.Context(), co.devfileMetadata.componentType)
-	}
-	err = co.devfileRun(cmd)
-	if err != nil {
-		return err
-	}
 	if log.IsJSON() {
 		return co.DevfileJSON()
 	}
+
 	return nil
 }
 
@@ -692,4 +393,28 @@ func NewCmdCreate(name, fullName string) *cobra.Command {
 	completion.RegisterCommandFlagHandler(componentCreateCmd, "binary", completion.FileCompletionHandler)
 
 	return componentCreateCmd
+}
+
+func getContext(now bool, cmd *cobra.Command) (*genericclioptions.Context, error) {
+	if now {
+		return genericclioptions.NewContextCreatingAppIfNeeded(cmd)
+	}
+	return genericclioptions.NewOfflineContext(cmd)
+}
+
+func getEnvFilePath(componentContext string) string {
+	if componentContext != "" {
+		return filepath.Join(componentContext, EnvYAMLFilePath)
+	}
+	return filepath.Join(LocalDirectoryDefaultLocation, EnvYAMLFilePath)
+}
+
+// DevfileParseFromFile reads, parses and validates a devfile from a file without flattening it
+func devfileParseFromFile(devfilePath string, resolved bool) (parser.DevfileObj, error) {
+	devObj, _, err := devfile.ParseDevfileAndValidate(parser.ParserArgs{Path: devfilePath, FlattenedDevfile: &resolved})
+	if err != nil {
+		return parser.DevfileObj{}, err
+	}
+
+	return devObj, nil
 }
